refactor: rename .native to platformless format, use .web for the web
