+++ src/lib/db.web.ts
@@ -1,0 +1,1 @@
+export const db = {updateHook: () => {}};
