+++ src/lib/queryClient.ts
@@ -1,6 +1,6 @@
 import {QueryClient} from '@tanstack/react-query';
 import {setupDatabaseHook} from './setupDatabaseHook';
-import {db} from './db.native';
+import {db} from './db';
 
 export const queryClient = new QueryClient();
