+++ src/mutations/messages.web.ts
@@ -1,0 +1,5 @@
+// import {db} from '../lib/db';
+
+export const createMessage = () => {
+  return Promise.resolve();
+};
