+++ src/queries/kv.ts
@@ -1,5 +1,5 @@
 import {useSuspenseQuery} from '@tanstack/react-query';
-import {db} from '../lib/db.native';
+import {db} from '../lib/db';
 import type {Chat} from '../lib/types';
 
 export const getByKey = async <T>(key: string) => {
